app/requirements.txt
flask==3.1.2
gunicorn==23.0.0
prometheus_client==0.22.1
psycopg2-binary==2.9.10
SQLAlchemy==2.0.43
python-dotenv==1.1.1

